Use consistent return codes in PubSubClient.
Some methods previously returned 0 for success, (and for failure) and some returned 1 for success and 0 for failure.  Make them all use 0 for success, and different error codes for each failure, and make them proper defines.

This is BACKWARDS INCOMPATIBLE unfortunately :(

Calls that were
   while (client.connect(config.devicename) != 1)
Should now be
   while ((rc = client.connect(config.devicename)) != ERR_OK)

The old code used 1 for success, 0 for failure for:
   connect()
   loop()

The old code used 0 for both success and failure for:
   publish()

The old code used no return code at all for:
   subscribe()

In all cases, the new code returns 0 (ERR_OK) for success, and either
ERR_NOT_CONNECTED or ERR_TIMEOUT_EXCEEDED for the failure cases.  They are now
also broken out per failure code.