Add classes implementing all of the MQTT message types, use them in PubSubClient
This feels a little like I've reimplmented PubSubClient.
I wanted to support more options for the various messages and that meant creating a separate class just for message data.
Then to handle the different options for the different messages, I had to create a class for each message type.

Streaming functionality is currently not implemented. It is commented out for now.