Fix building for non-ESP8266 devices
** API BREAKAGE **

We have to go back to passing a "Client" object to the PubSubClient constructor. I didn't like this when I originally saw it and moved the WiFiClient object into the class. But this meant that other platforms couldn't easily provide compatible objects from their own network stacks.
I have no idea if this work on the Arduino Ethernet/WiFi/Yun boards because I don't have them. It compiles though.

Brought to light by #16.