Fix reading multi-packet messages in MQTT::readPacket()
The pointer passed to client.read() wasn't being advanced, so the loop was overwriting the start of the buffer.

Thanks to Mike Hole and his bug report over on Azure (Azure/azure-iot-sdks#393) for providing enough clues to find this mistake. I guess all of my testing so far has involved single packet messages.

Also, put code in a block to limit variable scope.