Added reference to PubSubClient object to callback prototype
As it is, the callback function relies on an external/global instance of
the PubSubClient object if it wants to use the object for instance for
publishing.

It is preferable to be able to pass a reference to the PubSubClient
instance into the callback function when it is called.

To that end the callback signature and the calls to the callback in the
loop method had to be changed.