PubSubClient: Add accessor functions to monitor Tx/Rx byte metrics
That this provides getTxCount() and getRxCount() which return
the number of bytes transmitted to and received from the underlying
stream by the MQTT implementation.

Note that this does not include any additional TCP headers that
may be added by the underlying implementation but it does give
a useful starting point for datacomms. usage over time e.g. when
using a cellular stream implementation such as TinyGSM

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>